+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Version 0.9 (beta-9)
 
 - Stricter naming of prohibited items
+- Detail the error when a label name containing `@` is undefined
 
 ## Version 0.8 (beta-8)